Improve UI of Editor and refactor code
In Editor.js, improved user interface by adding the function to move forward and backward every 5 seconds using the icons TbRewindBackward5 and TbRewindForward5, respectively. Also, it added the functionalities to go to start and end of the video using the icons IoPlayBackOutline and IoPlayForwardOutline, respectively. The video now can be paused or played by hovering the mouse on the video, with a fade-in/out transition. The imported SVG file tryOut.svg is used as an icon for the upload video button.

The code was also refactored for readability: indentations were adjusted, unused console logs were removed, and additional information about the video was added to the videoMetadata.

The changes were made to enhance user convenience and maintain code legibility.