Add audio detection in video editor
Added a new feature in the 'src/pages/Editor.js' file, allowing the application to detect whether a video has audio or not. This was done by defining a function, 'hasAudio', that checks for audio in the metadata of the video file. An event listener was added to output a confirmation message when the video data is loaded. This update enhances user experience by providing real-time feedback about the presence of audio in the loaded video.