Improve audio detection in video editor
Refactored the way the application determines whether a video has audio. Instead of using properties like 'mozHasAudio' or 'webkitAudioDecodedByteCount', we now use an analyser node from the Web Audio API to check if there are any frequencies present in the audio track. This method is more reliable and supports more browsers. Also, changed how the video behaves when it does have or doesn't have audio. Now we pause the video and revoke its URL when it doesn't have audio.